fix(connectors): send the body in the encoding the spec declares
A real Mailgun account, a real sending key: posting the form Mailgun asks
for returns 200 and the mail arrives. Posting the same fields as JSON to
the same endpoint with the same key returns 400 "from parameter is
missing" — it does not object to the format, it simply cannot see the
fields. The runtime only ever sent JSON, and the parser only ever read
the application/json schema, so the media type a spec declares was never
looked at. Mailgun, Twilio and Stripe all take form-encoded bodies, so
"assemble a connector from the vendor's own spec" was broken for a large
part of the real world.

Read the media type from the spec and encode to it. JSON stays the
default and stays first when a spec offers both, so existing connectors
are untouched. multipart says plainly that it is unsupported rather than
falling back — sending the wrong encoding reads, on the far side, as
"you left these fields out", which sends whoever debugs it to inspect a
binding that is perfectly correct.

The stand-in learns to want a form first: one route now accepts only
form fields and answers a JSON body with Mailgun's own sentence. Every
fake vendor in the mock had spoken JSON, which is why no test had ever
met an API that does not.

Carries the spec size ceiling with it: it was a compile-time constant
copied into the frontend, and GitHub publishes a 12 MB description, so
an owner had nothing to turn. It is CONNECTOR_SPEC_MAX_BYTES now, the
frontend copy is gone, and the default is unchanged.

+++ backend/internal/connector/svc_validate.go
@@ -94,7 +94,7 @@ func (s *Service) fetchSpec(ctx context.Context, url string) ([]byte, error) {
 // 真世界里这不是边角：GitHub 自己发布的 `api.github.com.json` 是 12 MB。
 // 粘贴那条路一直说得对（`ValidateIngest` 先量长度），是这条抓取的路够不着同一句话。
 func readSpecResponse(resp *http.Response) ([]byte, error) {
-	raw, rerr := io.ReadAll(io.LimitReader(resp.Body, MaxSpecBytes+1))
+	raw, rerr := io.ReadAll(io.LimitReader(resp.Body, int64(MaxSpecBytes)+1))
 	if cerr := resp.Body.Close(); cerr != nil && rerr == nil {
 		rerr = cerr
 	}
